refactor: move getDataTableCostField function to costAnalysis helper
