+++ apps/web/src/services/cost-explorer/cost-analysis/modules/CostAnalysisDataTable.vue
@@ -33,6 +33,7 @@ import { arrayToQueryString, objectToQueryString, primitiveToQueryString } from 
 import ErrorHandler from '@/common/composables/error/errorHandler';
 
 import { ASSET_INVENTORY_ROUTE } from '@/services/asset-inventory/route-config';
+import { getDataTableCostFields } from '@/services/cost-explorer/cost-analysis/lib/data-table-helper';
 import { DATE_FORMAT } from '@/services/cost-explorer/cost-analysis/lib/widget-data-helper';
 import UsageTypeAdditionalFilterSelector from '@/services/cost-explorer/cost-analysis/modules/UsageTypeAdditionalFilterSelector.vue';
 import type { UsageTypeAdditionalFilter } from '@/services/cost-explorer/lib/config';
@@ -45,7 +46,6 @@ import {
     USAGE_TYPE_ADDITIONAL_FILTER_MAP,
 } from '@/services/cost-explorer/lib/config';
 import {
-    getDataTableCostFields,
     getTimeUnitByGranularity,
 } from '@/services/cost-explorer/lib/helper';
 import { useCostAnalysisPageStore } from '@/services/cost-explorer/store/cost-analysis-page-store';
